CryptoToken and Keypair Cache. No more hardcoded Token credentials
Created a cache in class Ed25519 that keeps in memory information on the cryptotoken including  keypairs, session info, token name, authentication code, slot number. The cache is filled on startup when the token is activated and only deals with the Ed25519 keypairs, leaving ejbca normal functioning for the others. Only one session is kept active and in cache per token preventing hanging sessions. Needs further work when editing changes to a Token are made by ejbca gui.